+++ src/doom/r_segs.cpp
@@ -784,7 +784,7 @@ void R_StoreWallRange(int start,
     }
 
     // calculate rw_offset (only needed for textured lines)
-    segtextured = midtexture | toptexture | bottomtexture | maskedtexture;
+    segtextured = midtexture | toptexture | bottomtexture | static_cast<int>(maskedtexture);
 
     if (segtextured)
     {
