+++ src/heretic/r_segs.cpp
@@ -558,7 +558,7 @@ void R_StoreWallRange(int start, int stop)
 //
 // calculate rw_offset (only needed for textured lines)
 //
-    segtextured = midtexture | toptexture | bottomtexture | maskedtexture;
+    segtextured = midtexture | toptexture | bottomtexture | static_cast<int>(maskedtexture);
 
     if (segtextured)
     {
