+++ src/hexen/r_segs.cpp
@@ -536,7 +536,7 @@ void R_StoreWallRange(int start, int stop)
 //
 // calculate rw_offset (only needed for textured lines)
 //
-    segtextured = midtexture | toptexture | bottomtexture | maskedtexture;
+    segtextured = midtexture | toptexture | bottomtexture | static_cast<int>(maskedtexture);
 
     if (segtextured)
     {
